Suggestion password helper
In the recent time I created helper to the password hashing and verify.
My class is smaller and more legible than current. Could somebody look at it?

repo: https://github.com/QsmaPL/Simple-PHP-Password-hasher

It is possible so that it is to add it to SMVCF and to replace old class? :)