Proposal to deal with multilanguage
Hello,

I've seen how the framework deals with "languages", but I think the module was poorly implemented. Ok, I know that you can change the language by accessing the language controller and the change() method. But the module doesn't seem complete. What I mean by this is that the language slug ("en", "fr", "ro") should find itself inside the url, because if not, I think the search engines will consider that the content served on the same url is different, as the same urls will serve "same content" but in different languages. I sure hope you understand what I mean here...

Now, in order to prevent this from happening we should have the language slug appended to the url, if the language is not the "default" one. I tried to deal with this inside the Url helper, but it doesn't seem to work, as the defined routes would bypass the parsing of the url.

The only viable solution was inside the Router.php.

Now, if the framework has more than one language defined, the dispatch() method will look if the slug is inside the url, if the slug is passed inside the url, then the language is set inside a session and a cookie.

If inside the url there is no language defined, the dispatch() method will look for the language inside the cookie and the session and if it finds out that the language is different than the default language, then it will redirect to an url that has the language slug appended.

This way of dealing with the url would also permit us to simply eliminate the language controller, as everything regarding the language would be dealt with inside the router.

Hoping that this would be aproved...